Better search functionality
Better state mangement, a couple bug fixes and enhancements. Moved dependencies to an ActivityComponent for better access to the single activity.

+++ Movies/moviedb/src/main/java/com/dev/cameronc/moviedb/data/MovieCreditsResponse.kt
@@ -13,7 +13,7 @@ data class MovieCreditsResponse(
 
 data class Cast(
         @SerializedName("cast_id")
-        val castId: Int,
+        val castId: Long,
         @SerializedName("character")
         val character: String,
         @SerializedName("credit_id")
